+++ __init__.py
@@ -5,7 +5,7 @@ from . import cfg
 from . import add_node_group_to_storage
 from . import update_node_group
 from . import get_node_group_from_storage
-from . import BIS_getNodesFromStorage
+from . import get_nodes_from_storage
 from . import nodes_panel
 from . import BIS_addTextToStorage
 from . import BISUpdateText
@@ -19,7 +19,8 @@ from . import update_mesh_in_storage
 from . import get_mesh_from_storage
 from . import WebRequests
 from . import MessageBox
-from . import BIS_Items
+from . import bis_items
+from . import nodes_tools_ops
 if cfg.experimental_enable_bis_custom_nodes:
     from . import nodes_bis_custom
 
@@ -28,7 +29,7 @@ bl_info = {
     'name': 'BIS',
     'category': 'Material',
     'author': 'Nikita Akimov',
-    'version': (1, 5, 1),
+    'version': (1, 5, 2),
     'blender': (2, 79, 0),
     'location': 'T-Panel > BIS',
     'wiki_url': 'https://b3d.interplanety.org/en/bis-online-blender-material-storage/',
@@ -41,7 +42,7 @@ def register():
     add_node_group_to_storage.register()
     update_node_group.register()
     get_node_group_from_storage.register()
-    BIS_getNodesFromStorage.register()
+    get_nodes_from_storage.register()
     nodes_panel.register()
     if cfg.experimental_enable_bis_custom_nodes:
         nodes_bis_custom.register()
@@ -57,11 +58,13 @@ def register():
     get_mesh_from_storage.register()
     WebRequests.register()
     MessageBox.register()
-    BIS_Items.register()
+    bis_items.register()
+    nodes_tools_ops.register()
 
 
 def unregister():
-    BIS_Items.unregister()
+    nodes_tools_ops.unregister()
+    bis_items.unregister()
     MessageBox.unregister()
     WebRequests.unregister()
     get_mesh_from_storage.unregister()
@@ -77,7 +80,7 @@ def unregister():
     if cfg.experimental_enable_bis_custom_nodes:
         nodes_bis_custom.unregister()
     nodes_panel.unregister()
-    BIS_getNodesFromStorage.unregister()
+    get_nodes_from_storage.unregister()
     get_node_group_from_storage.unregister()
     update_node_group.unregister()
     add_node_group_to_storage.unregister()