+++ node_manager.py
@@ -10,42 +10,44 @@ from .NodeNodeGroup import *
 from .node_node_group import *
 from .addon import Addon
 from .WebRequests import WebRequest
-from .BIS_Items import BIS_Items
+from .bis_items import BISItems
 
 
 class NodeManager:
 
     @staticmethod
-    def items_from_bis(context, search_filter, page, update_preview):
+    def items_from_bis(context, search_filter, page, update_preview=False):
         # get page of items list from BIS
         rez = None
+        storage_subtype = __class__.get_subtype(context)
+        storage_subtype2 = __class__.get_subtype2(context)
         request = WebRequest.send_request({
             'for': 'get_items',
             'search_filter': search_filter,
             'page': page,
             'storage': __class__.storage_type(context),
-            'storage_subtype': __class__.get_subtype(context),
-            'storage_subtype2': __class__.get_subtype2(context),
+            'storage_subtype': storage_subtype,
+            'storage_subtype2': storage_subtype2,
             'update_preview': update_preview
         })
         if request:
             request_rez = json.loads(request.text)
             rez = request_rez['stat']
             if request_rez['stat'] == 'OK':
-                preview_to_update = BIS_Items.updatePreviewsFromData(request_rez['data']['items'], __class__.storage_type(context))
+                preview_to_update = BISItems.update_previews_from_data(data=request_rez['data']['items'], list_name=__class__.storage_type(context))
                 if preview_to_update:
                     request = WebRequest.send_request({
                         'for': 'update_previews',
                         'preview_list': preview_to_update,
                         'storage': __class__.storage_type(context),
-                        'storage_subtype': __class__.get_subtype(context),
-                        'storage_subtype2': __class__.get_subtype2(context)
+                        'storage_subtype': storage_subtype,
+                        'storage_subtype2': storage_subtype2
                     })
                     if request:
                         previews_update_rez = json.loads(request.text)
                         if previews_update_rez['stat'] == 'OK':
-                            BIS_Items.updatePreviewsFromData(previews_update_rez['data']['items'], __class__.storage_type(context))
-                BIS_Items.createItemsList(request_rez['data']['items'], __class__.storage_type(context))
+                            BISItems.update_previews_from_data(data=previews_update_rez['data']['items'], list_name=__class__.storage_type(context))
+                BISItems.create_items_list(data=request_rez['data']['items'], list_name=__class__.storage_type(context))
                 context.window_manager.bis_get_nodes_info_from_storage_vars.current_page = page
                 context.window_manager.bis_get_nodes_info_from_storage_vars.current_page_status = request_rez['data']['status']
         return rez
@@ -92,12 +94,18 @@ class NodeManager:
     @staticmethod
     def get_subtype(context):
         # return subtype
-        return context.area.spaces.active.tree_type
+        if context.area.spaces.active.type == 'NODE_EDITOR':
+            return context.area.spaces.active.tree_type
+        else:
+            return 'ShaderNodeTree'
 
     @staticmethod
     def get_subtype2(context):
         # return subtype2
-        return context.area.spaces.active.shader_type
+        if context.area.spaces.active.type == 'NODE_EDITOR':
+            return context.area.spaces.active.shader_type
+        else:
+            return 'OBJECT'
 
     @staticmethod
     def is_procedural(material):